projects/api-python/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-JWT-Extended==3.12.1
requests==2.19.1
mysql-connector==2.1.7
pytest==3.7.4
pylint==2.1.1